Refactor packaging import boundaries
