+++ AGENTS.md
@@ -6,9 +6,9 @@ You are coding for the GEPA-Mindfulness-superalignment repository. Follow these 
 
 - Repository instructions are canonical here; beads reads them via `beads/README.md`.
 - See `beads/README.md` for beads setup and ingestion notes.
-- Reminder: Codex automation (agents/scripts) must avoid `bd`, while humans and onboarding
-  flows should continue to use `bd` for issue tracking and CLI workflows; direct `.beads/*`
-  edits are acceptable for artifact updates or when the CLI is unavailable.
+- Reminder: Codex automation for this repository must not invoke `bd` or edit `.beads/*`
+  unless the maintainer explicitly requests it. Human maintainers, onboarding flows, and
+  non-automated local workflows may continue to use `bd` for issue tracking and CLI workflows.
 - Beads is already enabled for this repository; run `bd onboard` from the repo root to fetch
   integration instructions. If `bd` is unavailable, ensure the beads CLI is on your PATH and
   ask the maintainer of the manual install for the expected location.
@@ -79,7 +79,10 @@ bd close bd-42 --reason "Completed" --json
 - `3` - Low (polish, optimization)
 - `4` - Backlog (future ideas)
 
-### Workflow for AI Agents
+### Workflow for Human-Mediated Agent Tasks
+
+These steps are for humans or explicitly authorized agent sessions. Codex automation should
+not run these commands unless the maintainer asks for beads work in the current task.
 
 1. **Check ready work**: `bd ready` shows unblocked issues.
 2. **Claim your task**: `bd update <id> --status in_progress`.
