--- src/__init__.py
@@ -1,1 +1,0 @@
-"""Helper package for CLI entry points."""
