+++ src/adversarial_evaluator.py
@@ -5,17 +5,18 @@ from __future__ import annotations
 import sys
 import warnings
 
-# NOTE: New implementation lives in dual_path_evaluator.py; keep this file as a thin shim.
+# NOTE: New implementation lives in mindful_trace_gepa.dual_path_evaluator.
 
 
 def main() -> int:
     warnings.warn(
-        "src.adversarial_evaluator is deprecated; use src.dual_path_evaluator instead.",
+        "src.adversarial_evaluator is deprecated; "
+        "use python -m mindful_trace_gepa.dual_path_evaluator instead.",
         DeprecationWarning,
         stacklevel=2,
     )
     try:
-        from src.dual_path_evaluator import main as dual_main
+        from mindful_trace_gepa.dual_path_evaluator import main as dual_main
     except ImportError as exc:
         print(f"Failed to import dual-path evaluator: {exc}", file=sys.stderr)
         return 1
